kubevirt: avoid dependency cycles
