+++ frontend/packages/kubevirt-plugin/src/components/modals/vm-description-modal/vm-description-modal.tsx
@@ -9,8 +9,9 @@ import {
   ModalComponentProps,
 } from '@console/internal/components/factory';
 import { k8sPatch } from '@console/internal/module/k8s';
+import { getDescription } from 'packages/kubevirt-plugin/src/selectors/selectors';
 import { VMLikeEntityKind } from '../../../types';
-import { getDescription, getVMLikeModel } from '../../../selectors/selectors';
+import { getVMLikeModel } from '../../../selectors/vm';
 import { getUpdateDescriptionPatches } from '../../../k8s/patches/vm/vm-patches';
 
 // TODO: should be moved under kubevirt-plugin/src/style.scss
