+++ frontend/packages/kubevirt-plugin/src/components/vm-disks/create-disk-row.tsx
@@ -13,8 +13,7 @@ import { StorageClassModel } from '@console/internal/models';
 import { getName } from '@console/shared';
 import { useSafetyFirst } from '@console/internal/components/safety-first';
 import { k8sPatch, K8sResourceKind } from '@console/internal/module/k8s';
-import { getVmPreferableDiskBus } from '../../selectors/vm';
-import { getVMLikeModel } from '../../selectors/selectors';
+import { getVmPreferableDiskBus, getVMLikeModel } from '../../selectors/vm';
 import { getAddDiskPatches } from '../../k8s/patches/vm/vm-disk-patches';
 import { VMLikeEntityKind } from '../../types';
 import { ValidationErrorType } from '../../utils/validations/common';
