+++ frontend/packages/kubevirt-plugin/src/components/vm-nics/create-nic-row.tsx
@@ -8,11 +8,10 @@ import { FormGroup, HelpBlock } from 'patternfly-react';
 import { useSafetyFirst } from '@console/internal/components/safety-first';
 import { k8sPatch, K8sResourceKind } from '@console/internal/module/k8s';
 import { getNamespace } from '@console/shared';
-import { getVMLikeModel } from '../../selectors/selectors';
 import { NetworkAttachmentDefinitionModel } from '../../models';
 import { getAddNicPatches } from '../../k8s/patches/vm/vm-nic-patches';
 import { VMKind, VMLikeEntityKind } from '../../types';
-import { getNetworkChoices } from '../../selectors/vm';
+import { getNetworkChoices, getVMLikeModel } from '../../selectors/vm';
 import { dimensifyRow } from '../../utils/table';
 import { NetworkType } from '../../constants/vm';
 import { getValidationErrorMessage, getValidationErrorType } from '../../utils/validations/common';
