+++ frontend/packages/kubevirt-plugin/src/components/vm-nics/vm-nics.tsx
@@ -7,8 +7,7 @@ import { useSafetyFirst } from '@console/internal/components/safety-first';
 import { sortable } from '@patternfly/react-table';
 import { createBasicLookup } from '@console/shared';
 import { VMLikeEntityKind } from '../../types';
-import { asVM } from '../../selectors/selectors';
-import { getInterfaces, getNetworks, getVmPreferableNicBus } from '../../selectors/vm';
+import { getInterfaces, getNetworks, getVmPreferableNicBus, asVM } from '../../selectors/vm';
 import { dimensifyHeader } from '../../utils/table';
 import { VMLikeEntityTabProps } from '../vms/types';
 import { NicRow } from './nic-row';
